+++ app/controllers/api/messages_controller.rb
@@ -1,6 +1,6 @@
 class Api::MessagesController < ApplicationController
   def random_greeting
     message = Message.order('RANDOM()').first
-    render json: { gretting: message.content }
+    render json: { greeting: message.content }
   end
 end
