+++ app/javascript/app.js
@@ -1,13 +1,22 @@
-import React from 'react';
-import {
-    BrowserRouter, Route,
-  } from 'react-router-dom';
-  import Greeting from './Greeting';
-  
-  const App = () => (
-    <BrowserRouter>
+import React, { useEffect } from 'react';
+import { useDispatch } from 'react-redux';
+import { Routes, Route } from 'react-router-dom';
+import Greeting from './Greeting';
+import { fetchMessage } from './redux/messageSlice';
+
+const App = () => {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    console.log('dispatch is working');
+    dispatch(fetchMessage());
+  }, []);
+
+  return (
+    <Routes>
       <Route path="/" element={<Greeting />} />
-    </BrowserRouter>
+    </Routes>
   );
-  
-  export default App;
+};
+
+export default App;
