+++ app/javascript/redux/messageSlice.js
@@ -1,18 +1,27 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
+import axios from 'axios'
+
 
 export const fetchMessage = createAsyncThunk('fetchMessage', async () => {
-  const response = await fetch(`/api/random_greeting`);
-  const data = await response.json();
-  return data;
+  const response = await axios.get('http://127.0.0.1:3000/api/random_greeting', {headers: {
+    Accept: 'application/json',
+  }});
+  return response.data
 });
 
+const initialState = {
+  value: {},
+  loading: false,
+  error : '',
+}
+
 const messagesSlice = createSlice({
-  name: 'message',
-  initialState: null,
+  name: 'messages',
+  initialState,
   reducers: {},
   extraReducers: (builder) => {
     builder
-      .addCase(fetchMessage.fulfilled, (state, action) => action.payload);
+      .addCase(fetchMessage.fulfilled, (state, action) => ({...state, value: action.payload}));
   },
 });
