+++ package.json
@@ -11,6 +11,7 @@
     "@hotwired/stimulus": "^3.2.2",
     "@hotwired/turbo-rails": "^7.3.0",
     "@reduxjs/toolkit": "^1.9.7",
+    "axios": "^1.5.1",
     "babel-loader": "^9.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
